transactify_service/api/webAPI/Customers.py
from store.webmodels.Customer import Customer
from store.serializers.CustomerSerializer import CustomerSerializer

from rest_framework import permissions
from rest_framework import renderers
from rest_framework.decorators import action
from rest_framework.response import Response
from rest_framework import viewsets

class CustomerViewSet(viewsets.ModelViewSet):
    """
    This ViewSet automatically provides `list`, `create`, `retrieve`,
    `update` and `destroy` actions.

    Additionally we also provide an extra `highlight` action.
    """
    queryset = Customer.objects.all()
    serializer_class = CustomerSerializer
    permission_classes = [permissions.IsAuthenticatedOrReadOnly]

    @action(detail=True, renderer_classes=[renderers.StaticHTMLRenderer])
    def highlight(self, request, *args, **kwargs):
        customer = self.get_object()
        return Response(customer.highlighted)

    def perform_create(self, serializer):
        serializer.save(owner=self.request.user)